Search: Replace fetch_input_blocking with fetch_input
The fetch_input_blocking and fetch_input functions are very much the
same function with the exception being fetch_input be a async function.
The original purpose for the function was to continue showing output,
while the user types the search query. This didn't work as we it should
have been, so they are now merged into a single function