Pager::push_str: fix bug where each call is a newline
Pager::push_str considered each call as a new line, regardless of
whether the last pushed line was terminated or not.

This commit fixes this behaviour, it now checks the last push,
whether it contains a newline or not. If it does, it moves it to
`wrap_lines`. If it doesn't, it pushes it to the `lines`

In case, it receives a text containing newlines in between, it moves
them individually to the `wrap_lines` and leaves the last line, if it
isn't \n terminated

The std::fmt::Write impl now directly uses `push_str`