async: Drop async support in minus in favour of threads
Async tasks are meant for IO bound operations which return after a specific
amount of time. minus is not one of those tasks because it may never
return as it is ultimately dependend on the user for doing it. Hence
this patch now uses threads for starting the pager. Threads are better
suited for long running tasks and offer much better speed

Therefore this patch switches the internal architecture of minu to
depend on threads rather than async runtimes

Following changes were introduced
* Remove async_paging feature in favour of dynamic_paging
* Replace it with dynamic_pager.rs
* Remove reader functions in favour of only polling based reader
* Add event_reader function to read user events
* Build failures due to api changes

There might be examples/doctests/tests failures and unupdated
documentation after this. They will be fixed/updated in follow up
commits