AppendStr: No longer redraws the entire screen
Whenever minus encountered a AppendStr event, it appended the given
string data into it's vector of lines and rdraw the entire screen. This
was the behaviour even before AppendStr when push_str function was used
internally: basically from day one.
This was a large problem in performance as writing to a terminal is
really slow and redrawing the entire screen every time a AppendStr event
occured was a huge bottleneck in terms of performance. Also because of this
behaviour minus also needed to keep track of some variables like filled
which were only required because of this way of appending data to the screen.

This patch just fixes the problem exactly as one expects
* Format the given string data and calculate the number of rows the data
  requires.
* Move the cursor to the next line of the last displayed line in the
  terminal.
* Dump the formatted string. It will try to display the entire string if
  possible or partially if the length of the data exceeds the available
  the number of rows.

minus no longer need to keep track if terminal got filled because with this
approach terminal overflows will get handled automatically without any
special care.