init_core: Enabling static_output causes minus to quit
If both static_output and dynamic_output features are enabled and number
of lines are less than available rows than minus will display something
or nothing at all and quit even in dynamic paging. This was because we
only checked for feature flags in that check block.

This patch also does a runtime check before running those checks