make_append_str: Wrong unterminated calculation
This bug is caused by a text line which contains newlines in middle but
not at the end. For example, if this text is given

This is a line\nThis is another line\nThis is third line

In this scenario, the function will return 3 as the untermination, when
in reality it should have been 1

This patch fixes this by making it a separate case and handling it
separately. It first formats all the terminated lines and then formats
the last line seprately to calculate the unterminated only for the last line.