make_append_str: Fix wrong calculation of unterminated
When sending complicated sets of data to this function, a wrong value of
unterminated was calculated. This caused really weird displaying and
truncation of PagerState::formatted_lines.

The reason for this miscalculations are way beyond describing due to the
over complexity of the make_append_str function. This patch simplifies
the entire function and causes correct unterminated value to be produced
wven when really complex data is given to the pager.

Fixes #66