feat: Introduce native input matching definition
This set of changes will allow minus to define key, mouse and resize
events to be declared in natively and not requiring the end application
to interact with the underlying crossterm crate.

This means that keybindings like Ctrl+L can be declared into minus with
appropriate function programatically using something like this

            pager.add_input_event(KeyBind("c-l", |pagerstate| {
                ...
            }));

This has the advantage that end-applications now get the ability to
add to, modify, or remove minus's default event binding without needing
to rewrite the entire matcher function by themselves.

This is the first commit to this major change and this adds the
necessory core functions and declarations to parse the keybindings.
Parsing of keybindings is fairly complete but extensive tests are
missing for these functions. I will add them in upcoming commits