feat: Introduce native input matching definition
This set of changes will allow minus to define key, mouse and resize
events to be declared in natively and not requiring the end application
to interact with the underlying crossterm crate.

This means that keybindings like Ctrl+L can be declared into minus with
appropriate function programatically using something like this

            pager.add_key_event("c-l", |event, pagerstate| {
                ...
            }));

This gives advantage to the end-applications as now they get the ability
to add to, modify, or remove minus's default event binding without needing
to rewrite the entire matcher function by themselves.