init_core: Handle runtime panics better (#77)
* init_core: Handle runtime panics better

In previous versions, during a panic the error message (and backtrace
possibly) might not get displayed after a panic. This happened because
they were printed on the alternate screen and after the program
exited, the terminal automatically switches to the main screen. hence
the message never gets to be displayed.

This patch fixes this by doing a cleanup before proceeding further on
the unwinding of the panic

* init: Fix panic handler not displaying message

No panic message would get displayed because of this bug. This occurred
because we take out the panic hook insside the panic::set_hook function

* init: Do not create noisy panics

With the new panic handler, a lot of noise was created for poisoned locks
and thread panics. This patch introduces checks for poisoned locks and
handles scoped thread panics as MinusError::Panic

* Revert "init: Do not create noisy panics"

This reverts commit 1f88576a65480eed194914e786e5693503d8f0d3.

* init: Do not propagate panics as Rust errors

This patch is introduced as according to Rust standards panics are
unrecoverable errors so they should propagate as it is