display: Intelligent refresh for movement
The following set of changes will enable minus to calculate what lines
need to be redrawn after a user scrolls up or down and only change those
lines.

This means that there will be no complete redraws of the terminal unless
it's necessary. This will cause much faster display output for larger
terminal windows and monitors and also improves the overall performance
of minus

Here are all the changes in chronological order

- Completed implementing the next drawing function
- Integrated draw_for_change in ev_handler
- Use new display function for static output
- Write docs for draw_on_change
- Add tests for draw_on_change