display: Do less refresh on single scrill downs
The following set of changes will enable minus to calculate what lines
need to be redrawn after a user scrolls up or down and only change those
lines.

This means that there will be no complete redraws of the terminal unless
it's necessory. This will cause much faster display output for larger
terminal windows and monitors and also improves the overall performance
of minus