display: Improve perf and fix a minor bug
The if statment in Ordering::Greater branch in draw_for_change function
checked whether normalized_delta is less than writable_rows, which is
always true and it should have been the delta variable which should be
checked. This patch fixes it.

When delta is greater than writable_rows, we now simply clear the entire
screen and start writing from the top row. This is much simpler than the
previous code that was present. This also improves performance by a bit.