search: Fix character loss while searching
After start or completion of search query prompt, certain key press
inputs were lost. This was because of the std::hint::spin_lock()
and AtomicBool that minus used for keeping track of whether main input
thread is running. These were chosen to get the most optimum performance
but they turned out to be faulty for what minus required.

This patch moves to parking_lot's Condvar and Mutex types to replace the
very same functionality and giving the same level of performance.