Don't paint after processing exit events
Previously, if the pager was in ExitStrategy::PagerQuit, it would paint
the pager one last time after receiving an exit event from the user,
leaving junk on the screen.

As far as I can tell, the conditional changed in this patch was intended
to avoid this extra paint, but had flipped &&/||, leading to this bug.