fix(display): Broken output due to mouse scroll
In a situation where there is incoming data and the user is
simultaneously scrolling through the text using the mouse causes the
output to get broken. Essentially there would be a few blank lines
that popped between two consecutive lines causing the broken display.

This occured because the value of `delta` in the `draw_for_change()`
was always equal to the net change in the `new_upper_mark` and the
previous upper_mark value regardless of whether the new lower mark value
goes beyond the available number of lines. Hence when the screen is
scrolled up using the `ScrollUp` crossterm command, the screen would
display blank lines for all the lines for which no text is returned by
the `PagerState::get_flattened_lines_with_bounds`.

The function does have checks to ensure new lower mark is never greater
than the number of lines. If this situation occurs, it sets the
new upper mark to number of lines - available rows. But this bug occured
because delta was being calculated before these checks.

This patch fixes it by moving the code that calculates the delta after
the checks

Closes #92