fix: Lines break where length pf line num increase
When the length of line count increases, say 2 at 99 to 3 at 100 the
line gets broken and becomes duplicated, if line 100 isn't terminated
with `\n` initially.
This bug occurs when line 100 is completed in more than one step,
something line:-
  Line 99 part 1
  Line 99 part 2\n
  Line 100 part 1
  Line 100 part 2
  Line 100 part 3\n

The output would be:-
  Line 99 part 1 Line 99 part 2\n
  Line 100 part 1
  Line 100 part 1 Line 100 part 2 Line 100 part 3\n

When the output should have been
  Line 99 part 1 Line 99 part 2\n
  Line 100 part 1 Line 100 part 2 Line 100 part 3\n

This occured because minus didn't keep a track of unterminated lines
when the length of line number increased.

This patch fixes the issue by assigning `p.unterminated` to the
unterminated lines when length of line number changes

Resolves #66