bugfix(input): `[n]G` takes to nth row not nth line
When pressing `[n]G`, minus would take the user to the nth row on the
screen and not the actual nth line. This occured because the `G` command
was initially designed by keeping in mind that one line occupied exactly
one row on the screen. With furthur development of minus this is no
longer an assumption and one line can occupy several rows on the
terminal but this implemention of `G` was never updated accordingly to
handle such cases.

This patch fixes the issue by introducing a new hash map
`lines_to_row_map` inside `PagerState` which describes on which row of
`PagerState::formatted_lines`, the first row of each line is present.
Finally the implementation of `[n]G` matches n in the map to get the row
and jumps the screen to that row.