fix(init): Text from main screen cleared upon quitting
This bug caused text from main screen being cleared upon exitting minus.
The reason for this is that `dynamic_paging` and `page_all` functions
prematurely set the RUNMODE mode which causes functions to act as if the
terminal is already setup with alternate screen active.

To fix this, we move the code that sets the RUNMODE variable inside
`init_core` function and ask the `dynamic_paging` and `page_all` functions
to pass the value of that variable to `init_core` as an argument.